#490 Extract Wallet interface and related changes
